Refactor external_control into a package
